requirements.txt
sys
cv2
glob
pandas
cryptography.fernet
uuid}